+++ fidget/src/core/vm/mod.rs
@@ -216,6 +216,10 @@ impl<const N: usize> Function for GenericVmFunction<N> {
     fn vars(&self) -> &VarMap {
         &self.0.vars
     }
+
+    fn can_simplify(&self) -> bool {
+        self.0.choice_count() > 0
+    }
 }
 
 impl<const N: usize> RenderHints for GenericVmFunction<N> {